+++ vf_base/__resource.lua
@@ -1,11 +1,10 @@
 resource_manifest_version '44febabe-d386-4d18-afbe-5e627f4af937'
 resource_type 'gametype' { name = 'venomous-freemode' }
-resource_version '1.1'
+resource_version '1.1.1'
 
-dependencies {'ghmattimysql', 'NativeUI'}
+dependencies {'ghmattimysql'}
 
 client_scripts {
-    '@NativeUI/NativeUI.lua',
     'config/freemode.lua',
     'config/spawn.lua',
     'config/vehicles.lua',