+++ vf_interaction/__resource.lua
@@ -1,4 +1,5 @@
 resource_manifest_version '44febabe-d386-4d18-afbe-5e627f4af937'
+dependency 'NativeUI'
 
 client_scripts {
     '@NativeUI/NativeUI.lua',