+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 1.1.1
 
 - Fixed death screen loop when the player changed his skin.
+- Removed NativeUI from vf_base.
 
 ## 1.1